requirements.txt
mysql-connector-python==8.0.30
fastapi==0.81.0
pydantic==1.9.2
mangum==0.15.1
python-dotenv==0.20.0
boto3==1.24.86
botocore==1.27.86
